+++ Runtime/RHI/GraphicsDriver.h
@@ -168,7 +168,7 @@ namespace Sailor::RHI
 			const TVector<RHI::RHITexturePtr>& colorAttachments,
 			RHI::RHITexturePtr depthStencilAttachment,
 			glm::ivec4 renderArea,
-			glm::vec2 offset,
+			glm::ivec2 offset,
 			glm::vec4 clearColor) = 0;
 
 		SAILOR_API virtual void EndRenderPass(RHI::RHICommandListPtr cmd) = 0;
